Split out MPIJobController into a separate module (prep for Register merge).
