+++ libensemble/tests/controller_tests/test_jobcontroller_multi.py
@@ -18,7 +18,8 @@ def build_simfunc():
 #--------------- Calling script ---------------------------------------------------------------
 
 from libensemble.register import Register, BalsamRegister
-from libensemble.controller import JobController, MPIJobController
+from libensemble.controller import JobController
+from libensemble.mpi_controller import MPIJobController
 from libensemble.baslam_controller import BalsamJobController
 
 #sim_app = 'simdir/my_simjob.x'
